Use Entity.ts organization pattern over index.ts
